+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Support triggers migrated from MySQL 5.6 to 5.7 in mysqldef [#157](https://github.com/k0kubun/sqldef/pull/157)
 - Fix duplicated `;`s of triggers in `mysqldef --export`
 
 ## v0.11.8